finished the jquery slide
finished the jquery slide. Used the addClass and removeClass to modify the .css(z-indez) from positive to negative.  Used the .next() to specify that the let nextSlide would be the current.next();
also used the .first() and .last() to specify that when the .length  ==0 the   activeslide = to the = first()/last()